Make req.cmd be 16 bits wide instead of 32
It's 16 bits wide on the network, and the code properly assumes so.
There's no need for it to take 32 bits in the structure (and it's
misleading too).

+++ nmdb/req.h
@@ -24,7 +24,7 @@ struct req_info {
 
 	/* operation information */
 	uint32_t id;
-	uint32_t cmd;
+	uint16_t cmd;
 	uint16_t flags;
 	const unsigned char *payload;
 	size_t psize;
